Put templates under 1 item
