+++ packages/app/src/app/pages/Dashboard/Content/routes/Templates/BookmarkedTemplates/mutations.ts
