+++ packages/app/src/app/pages/Dashboard/Content/routes/Templates/Navigation/elements.ts
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import { Link } from 'react-router-dom';
 
 export const Container = styled.div`
   position: relative;
@@ -7,14 +6,17 @@ export const Container = styled.div`
   align-items: center;
 `;
 
-export const NavigationLink = styled(Link)`
+export const NavigationTitle = styled.h2`
+  display: inline-flex;
   margin-right: 0.5rem;
-  color: rgba(255, 255, 255, 0.6);
   text-decoration: none;
   transition: 0.3s ease color;
   font-size: 1.25rem;
   margin-left: 0;
   color: white;
+  margin-top: 0;
+  margin-bottom: 0;
+  font-weight: 400;
 
   &:last-child {
     margin-right: 0;
