+++ packages/app/src/app/pages/Dashboard/Content/routes/Templates/OwnedTemplates/index.ts
@@ -1,0 +1,1 @@
+export { OwnedTemplates } from './OwnedTemplates';
