+++ packages/app/src/app/pages/Dashboard/Content/routes/Templates/index.ts
@@ -1,2 +1,1 @@
-export { Templates } from './Mine/index';
-export { FollowedTemplates } from './FollowedTemplates/index';
+export { Templates } from './Templates';
