--- packages/app/src/app/pages/Dashboard/Content/routes/Templates/types.ts
@@ -1,5 +1,0 @@
-export interface IMatch {
-  params: {
-    teamId: string | undefined;
-  };
-}
